Chunk 10: split SystemDashboard.jsx (M2)
Extraction-only, no behaviour change:
- systemDashboardConstants.js: EMPTY_FORM, SECTIONS, getVal helper
- CreateTenantForm.jsx: new-tenant form section
- RestoreBackupSection.jsx: restore-from-backup section
- RestoreConfirmModal.jsx: destructive-restore confirmation modal
- FeatureConfigModal.jsx: per-tenant feature-toggle modal

SystemDashboard.jsx 832 -> 458 lines.
